+++ angularwithasp.client/src/app/components/garage-list/garage-list.component.ts
@@ -1,31 +1,43 @@
 import { Component, OnInit } from '@angular/core';
 import { GarageService } from '../../services/garage.service';
 import { Garage } from '../../models/garage.model';
+import { CommonModule } from '@angular/common';
+import { Router } from '@angular/router';
 
 @Component({
-  selector: 'app-garage-list',
-  standalone: false,
-  templateUrl: './garage-list.component.html',
-  styleUrls: ['./garage-list.component.scss']
+    selector: 'app-garage-list',
+    templateUrl: './garage-list.component.html',
+    styleUrls: ['./garage-list.component.scss'],
+    imports: [CommonModule]
 })
 export class GarageListComponent implements OnInit {
-  garages: Garage[] = [];
+    garages: Garage[] = [];
 
-  constructor(private garageService: GarageService) { }
+    constructor(
+        private garageService: GarageService,
+        private router: Router
+    ) { }
 
-  ngOnInit(): void {
-    this.loadGarages();
-  }
+    ngOnInit(): void {
+        this.loadGarages();
+    }
 
-  loadGarages(): void {
-    this.garageService.getGarages().subscribe(garages => {
-      this.garages = garages;
-    });
-  }
+    loadGarages(): void {
+        this.garageService.getGarages().subscribe(garages => {
+            this.garages = garages;
+        });
+    }
 
-  deleteGarage(id: number): void {
-    this.garageService.deleteGarage(id).subscribe(() => {
-      this.loadGarages();
-    });
-  }
+
+    deleteGarage(id: number): void {
+        if (confirm('Êtes-vous sûr de vouloir supprimer ce garage ?')) {
+            this.garageService.deleteGarage(id).subscribe(() => {
+                this.loadGarages();
+            });
+        }
+    }
+
+    updateGarage(id: number): void {
+        this.router.navigate(['garage-edit', id]);
+    }
 }
